authentication-app/requirements.txt
pyserial
#Django==4.2.6
redis
flask==2.3.3
Flask-Session[redis]
Flask-SQLAlchemy==3.1.1
flask-bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
flask-wtf==1.2.1
flask-socketio
WTForms
Werkzeug==2.3.7
Flask-Assets==2.1.0
libsass